tests/wpt/tests/css/css-animations/parsing/animation-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation">
<meta name="assert" content="animation supports only the grammar '<single-animation> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("animation", "1s 2s 3s");
test_invalid_value("animation", "-1s -2s");

test_invalid_value("animation", "steps(1) steps(2)");

test_invalid_value("animation", "1 2");

test_invalid_value("animation", "reverse alternate alternate-reverse anim");

test_invalid_value("animation", "both backwards forwards anim");

test_invalid_value("animation", "paused running paused anim");

test_invalid_value("animation", "anim1 timeline1 anim2");
</script>
</body>
</html>
